Move OdomChassisControllerIntegrated

+++ include/okapi/api/chassis/controller/odomChassisController.hpp
@@ -28,7 +28,7 @@ class OdomChassisController : public virtual ChassisController {
    * @param imoveThreshold minimum length movement
    */
   OdomChassisController(std::shared_ptr<SkidSteerModel> imodel, std::unique_ptr<Odometry> iodometry,
-                        float imoveThreshold = 10);
+                        double imoveThreshold = 10);
 
   /**
    * Drives the robot straight to a point in the odom frame.
@@ -38,14 +38,14 @@ class OdomChassisController : public virtual ChassisController {
    * @param ibackwards whether to drive to the target point backwards
    * @param ioffset offset from target point in the direction pointing towards the robot
    */
-  virtual void driveToPoint(float ix, float iy, bool ibackwards = false, float ioffset = 0) = 0;
+  virtual void driveToPoint(double ix, double iy, bool ibackwards = false, double ioffset = 0) = 0;
 
   /**
    * Turns the robot to face an angle in the odom frame.
    *
    * @param iangle angle to turn to
    */
-  virtual void turnToAngle(float iangle) = 0;
+  virtual void turnToAngle(double iangle) = 0;
 
   /**
    * Passthrough to internal Odometry object.
@@ -67,10 +67,10 @@ class OdomChassisController : public virtual ChassisController {
    *
    * @param imoveThreshold new move threshold
    */
-  virtual void setMoveThreshold(float imoveThreshold);
+  virtual void setMoveThreshold(double imoveThreshold);
 
   protected:
-  float moveThreshold; // Minimum length movement
+  double moveThreshold; // Minimum length movement
   std::unique_ptr<Odometry> odom;
   CROSSPLATFORM_THREAD task;
 };
